requirements.txt
asttokens==3.0.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
seaborn==0.13.2
catboost==2.1.0
xgboost==2.1.0
